feat: 카카오 소셜 로그인 기본 템플릿 구현
- 카카오 인가 코드로 액세스 토큰 획득 및 프로필 조회
- 프로필 정보 기반 로그인/회원가입 분기 처리
- 미가입 사용자의 경우 AuthException를 통해 회원가입을 하도록 응답

TODO: 내부 사용자 테이블 생성 및 DB 연동 및 기능 구현 필요

+++ src/main/java/org/ringling/backend/auth/jwt/JavaJwtProvider.java
@@ -38,19 +38,19 @@ public class JavaJwtProvider {
             .build();
     }
 
-    public String createAccessToken(Long userId) {
+    public String createAccessToken(Integer userId) {
         return createAccessToken(userId, accessTokenExpirationMs);
     }
 
-    public String createAccessToken(Long userId, Long expirationMs) {
+    public String createAccessToken(Integer userId, Long expirationMs) {
         return createJWTToken(userId, TokenType.AccessToken, expirationMs);
     }
 
-    public String createRefreshToken(Long userId) {
+    public String createRefreshToken(Integer userId) {
         return createRefreshToken(userId, refreshTokenExpirationMs);
     }
 
-    public String createRefreshToken(Long userId, Long expirationMs) {
+    public String createRefreshToken(Integer userId, Long expirationMs) {
         return createJWTToken(userId, TokenType.RefreshToken, expirationMs);
     }
 
@@ -62,7 +62,7 @@ public class JavaJwtProvider {
      * @param expirationMs 만료 시간 (ms)
      * @return 서명된 JWT 문자열
      */
-    private String createJWTToken(Long userId, TokenType tokenType, Long expirationMs) {
+    private String createJWTToken(Integer userId, TokenType tokenType, Long expirationMs) {
         Instant now = Instant.now();
 
         return JWT.create()
